uhd: do not restart streaming on timing errors
A device streaming reset and restart was previously added to
handle a sample counter reset bug in a prerelease version of
UHD. The assumption was that in the event of a large timestamp
jump, forward or back, the error condition could be recovered
from by resetting the device clock. This recovery attempt is
not effective on more subtle timestamp errors and causes
additional errors with UHD. Remove the recovery mechanism
and fix the real problem with UHD timestamps.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>